+++ Source/OxyPlot.Wpf.Shared/PlotCommands.cs
@@ -1,6 +1,6 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="PlotCommands.cs" company="OxyPlot">
-//   Copyright (c) 2014 OxyPlot contributors
+//   Copyright (c) 2020 OxyPlot contributors
 // </copyright>
 // <summary>
 //   Provides a standard set of commands for the <see cref="PlotView" /> control.
@@ -19,6 +19,6 @@ namespace OxyPlot.Wpf
         /// <summary>
         /// Gets the value that represents the "Reset all axes" command.
         /// </summary>
-        public static readonly ICommand ResetAxes = new RoutedUICommand("Reset all axes", "ResetAxes", typeof(PlotBase));
+        public static readonly ICommand ResetAxes = new RoutedUICommand("Reset all axes", "ResetAxes", typeof(PlotViewBase));
     }
 }